MoE: extract compile_moe_reduction into moe_reduce.py
The topk-reduction kernel (compile_moe_reduction) lived in
moe_gemm_2stage.py alongside the int4 (a16wi4) stage1/stage2 GEMM
builders, but it is dtype-agnostic: it is the epilogue for stage2
mode="reduce" for every dtype (fp8/a8w4/mxfp4/int4), reached via
moe_kernels._run_moe_reduction. Move it to its own module so it lives
independently of the int4 GEMM builders.

- New self-contained kernels/moe_reduce.py (own _if_then/_if_else SCF
  helpers, no back-reference to moe_gemm_2stage).
- moe_kernels._run_moe_reduction now imports from .kernels.moe_reduce
  (the live production path).
- moe_gemm_2stage.py imports compile_moe_reduction back so the (unused)
  compile_moe_gemm2_ex still resolves; int4 gemm1/gemm2 untouched.

Cleanup on the moved kernel: drop the `if True:` scaffold, the unused
i8_type() helper, and two dead no-op statements (get_hip_arch(),
ir.ShapedType.get_dynamic_size()); trim step-narration comments while
keeping the load-bearing rationale (64-bit base-offset fold, unsigned
Index->ult, EP gather). Behavior-preserving: fx.Index, buffer_ops, and
the scf.IfOp control flow are kept as-is.

Verified on gfx950 (FLYDSL_RUNTIME_ENABLE_CACHE=0): _run_moe_reduction
matches a torch reference for f16/bf16/f32 x masked/unmasked, over both
the full-vector and tail (model_dim=4100) paths (max err <= 1.9e-6).
